🔧 update secret config
