+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/console.php
@@ -28,12 +28,6 @@ use Symfony\Bundle\FrameworkBundle\Command\DebugAutowiringCommand;
 use Symfony\Bundle\FrameworkBundle\Command\EventDispatcherDebugCommand;
 use Symfony\Bundle\FrameworkBundle\Command\RouterDebugCommand;
 use Symfony\Bundle\FrameworkBundle\Command\RouterMatchCommand;
-use Symfony\Bundle\FrameworkBundle\Command\SecretsDecryptToLocalCommand;
-use Symfony\Bundle\FrameworkBundle\Command\SecretsEncryptFromLocalCommand;
-use Symfony\Bundle\FrameworkBundle\Command\SecretsGenerateKeysCommand;
-use Symfony\Bundle\FrameworkBundle\Command\SecretsListCommand;
-use Symfony\Bundle\FrameworkBundle\Command\SecretsRemoveCommand;
-use Symfony\Bundle\FrameworkBundle\Command\SecretsSetCommand;
 use Symfony\Bundle\FrameworkBundle\Command\TranslationDebugCommand;
 use Symfony\Bundle\FrameworkBundle\Command\TranslationUpdateCommand;
 use Symfony\Bundle\FrameworkBundle\Command\WorkflowDumpCommand;
@@ -48,6 +42,12 @@ use Symfony\Component\Messenger\Command\FailedMessagesRetryCommand;
 use Symfony\Component\Messenger\Command\FailedMessagesShowCommand;
 use Symfony\Component\Messenger\Command\SetupTransportsCommand;
 use Symfony\Component\Messenger\Command\StopWorkersCommand;
+use Symfony\Component\Secret\Command\SecretsDecryptToLocalCommand;
+use Symfony\Component\Secret\Command\SecretsEncryptFromLocalCommand;
+use Symfony\Component\Secret\Command\SecretsGenerateKeysCommand;
+use Symfony\Component\Secret\Command\SecretsListCommand;
+use Symfony\Component\Secret\Command\SecretsRemoveCommand;
+use Symfony\Component\Secret\Command\SecretsSetCommand;
 use Symfony\Component\Translation\Command\TranslationPullCommand;
 use Symfony\Component\Translation\Command\TranslationPushCommand;
 use Symfony\Component\Translation\Command\XliffLintCommand;
