+++ src/Symfony/Component/Secret/AbstractVault.php
@@ -18,12 +18,11 @@ namespace Symfony\Component\Secret;
  */
 abstract class AbstractVault
 {
-    /** @var string|null */
     protected $lastMessage;
 
-    public function getLastMessage(): string
+    public function getLastMessage(): ?string
     {
-        return $this->lastMessage ?? '';
+        return $this->lastMessage;
     }
 
     abstract public function generateKeys(bool $override = false): bool;
