+++ src/Symfony/Component/Secret/CHANGELOG.md
@@ -1,0 +1,7 @@
+CHANGELOG
+=========
+
+6.1
+---
+
+* added the component
