+++ src/Symfony/Component/Secret/Command/SecretsSetCommand.php
@@ -13,7 +13,10 @@ declare(strict_types=1);
 
 namespace Symfony\Component\Secret\Command;
 
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Completion\CompletionInput;
+use Symfony\Component\Console\Completion\CompletionSuggestions;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -27,12 +30,11 @@ use Symfony\Component\Secret\AbstractVault;
  * @author Jérémy Derussé <jeremy@derusse.com>
  * @author Nicolas Grekas <p@tchwork.com>
  */
+#[AsCommand(name: 'secrets:set', description: 'Set a secret in the vault')]
 final class SecretsSetCommand extends Command
 {
-    protected static $defaultName = 'set';
-
-    private $vault;
-    private $localVault;
+    private AbstractVault $vault;
+    private ?AbstractVault $localVault;
 
     public function __construct(AbstractVault $vault, AbstractVault $localVault = null)
     {
@@ -42,15 +44,13 @@ final class SecretsSetCommand extends Command
         parent::__construct();
     }
 
-    protected function configure(): void
+    protected function configure()
     {
         $this
-            ->setDescription('Set a secret in the vault')
             ->addArgument('name', InputArgument::REQUIRED, 'The name of the secret')
             ->addArgument('file', InputArgument::OPTIONAL, 'A file where to read the secret from or "-" for reading from STDIN')
             ->addOption('local', 'l', InputOption::VALUE_NONE, 'Update the local vault.')
             ->addOption('random', 'r', InputOption::VALUE_OPTIONAL, 'Generate a random value.', false)
-            ->addOption('dir', 'd', InputOption::VALUE_REQUIRED, 'Target directory')
             ->setHelp(<<<'EOF'
 The <info>%command.name%</info> command stores a secret in the vault.
 
@@ -78,10 +78,9 @@ EOF
     {
         $errOutput = $output instanceof ConsoleOutputInterface ? $output->getErrorOutput() : $output;
         $io = new SymfonyStyle($input, $errOutput);
-
         $name = $input->getArgument('name');
         $vault = $input->getOption('local') ? $this->localVault : $this->vault;
-        $value = null;
+
         if (null === $vault) {
             $io->error('The local vault is disabled.');
 
@@ -121,11 +120,6 @@ EOF
             }
         }
 
-        if (false === is_string($value)) {
-            $io->warning('No value provided: using empty string');
-            $value = '';
-        }
-
         $vault->seal($name, $value);
 
         $io->success($vault->getLastMessage() ?? 'Secret was successfully stored in the vault.');
@@ -137,10 +131,17 @@ EOF
             $io->newLine();
         }
 
-        if ($this->vault === $vault && null !== $this->localVault && null !== $this->localVault->reveal($name)) {
+        if ($this->vault === $vault && null !== $this->localVault->reveal($name)) {
             $io->comment('Note that this secret is overridden in the local vault.');
         }
 
         return 0;
+    }
+
+    public function complete(CompletionInput $input, CompletionSuggestions $suggestions): void
+    {
+        if ($input->mustSuggestArgumentValuesFor('name')) {
+            $suggestions->suggestValues(array_keys($this->vault->list(false)));
+        }
     }
 }
