+++ src/Symfony/Component/Secret/composer.json
@@ -15,11 +15,14 @@
         }
     ],
     "require": {
-        "php": ">=8.1"
+        "php": ">=8.1",
+        "symfony/console": "^5.4|^6.0",
+        "symfony/dependency-injection": "^5.4|^6.0",
+        "symfony/dotenv": "^5.4|^6.0",
+        "symfony/var-exporter": "^5.4|^6.0"
     },
     "require-dev": {
-        "symfony/console": "^5.4|^6.0",
-        "symfony/dotenv": "^5.4|^6.0"
+        "symfony/filesystem": "^5.4|^6.0"
     },
     "autoload": {
         "psr-4": {
